requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
python-dotenv>=0.19.0